[WIP] First working implementation
Happy with the overalls scheme, makes object generation really simple from outside the engine and gives a lot of control to object manipulation.
The only issue that needs to be immediately resolved is the initial occlusion status of objects so that instances are probably initialized.

+++ MVCore/Systems/ActionSystem.cs
@@ -30,9 +30,9 @@ namespace MVCore.Systems
             ActionSceneNodes.Clear();
         }
 
-        public override void Update(float dt)
+        public override void Update(double dt)
         {
-            time += dt;
+            time += (float) dt;
 
             if (time < timeInterval)
                 return;
